// web/lib/api/generated/model/audienceTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ALL` - All shops
 * * `PRICE_LIST` - Shops on a price list
 * * `RETAILER` - One shop
 */
export type AudienceTypeEnum = typeof AudienceTypeEnum[keyof typeof AudienceTypeEnum];


export const AudienceTypeEnum = {
  ALL: 'ALL',
  PRICE_LIST: 'PRICE_LIST',
  RETAILER: 'RETAILER',
} as const;