// web/lib/api/generated/model/discountTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `PERCENT` - Percentage
 * * `FLAT_PER_UNIT` - Rupees off per unit
 */
export type DiscountTypeEnum = typeof DiscountTypeEnum[keyof typeof DiscountTypeEnum];


export const DiscountTypeEnum = {
  PERCENT: 'PERCENT',
  FLAT_PER_UNIT: 'FLAT_PER_UNIT',
} as const;