// web/lib/api/generated/model/scopeTypeEnum.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Inventory Platform API
 * Multi-tenant B2B inventory & ordering platform. All money values are decimal strings.
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `ALL` - All products
 * * `PRODUCT` - One product
 * * `CATEGORY` - A category (and its sub-categories)
 * * `BRAND` - A brand
 */
export type ScopeTypeEnum = typeof ScopeTypeEnum[keyof typeof ScopeTypeEnum];


export const ScopeTypeEnum = {
  ALL: 'ALL',
  PRODUCT: 'PRODUCT',
  CATEGORY: 'CATEGORY',
  BRAND: 'BRAND',
} as const;